backend/requirements.txt
asgiref==3.11.1
Django==4.2.28
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
mysqlclient==2.2.8
sqlparse==0.5.5
tzdata==2025.3
